Renamed deactivateFeatures as removeFeatures
