--- src/deactivate-features.js
@@ -1,3 +1,0 @@
-import difference from 'lodash/difference';
-
-export const deactivateFeatures = difference;
