+++ src/index.js
@@ -5,7 +5,7 @@ export { mergeFeatures } from './merge-features';
 export { isActive } from './is-active';
 export { FeatureToggles } from './feature-toggles';
 export { Feature } from './feature';
-export { deactivateFeatures } from './deactivate-features';
+export { deactivateFeatures } from './remove-features';
 export { parseQuery } from './parse-query';
 export { createRouteMiddleware } from './create-route-middleware';
 export { withFeatureToggles } from './with-feature-toggles';
