+++ src/remove-features.js
@@ -1,0 +1,3 @@
+import difference from 'lodash/difference';
+
+export const removeFeatures = difference;
